requirements.txt
graphene-django
graphene-file-upload
django_filter
django-graphene-permissions